Extract proto creation out of debug symbol parsing
Split existing debug symbol parsing in two steps. First, we fill a new
simple data structure, DebugSymbols. Then we have a centralized function
that takes care of converting the simple data to a protobuf. This is
done to better identify where time is spent when parsing symbols, and
also to make it easier to optimize the conversion to protobufs.
Also, add support for arena for symbols proto creation.