Fix "Too many directory blocks." issue in raw_pdb
The pdb validation step was modified so that we allow having more block
indices than what fits in a single block. The assumption here is that in
this case, the indices will spill in subsequent blocks. This was
verified for a single problematic pdb, but it's unclear if that would
always be the case. Also, we check that the indices blocks are
contiguous in memory so that we can safely index from the first block's
address. If this condition is not met, then we abort the pdb loading.