Make symbol loading more than 50x faster on Linux
Remove N^2 redundant sorting of functions data views when loading symbols.
On Linux, this makes Unreal symbols load in ~5 seconds instead of more
than 4 minutes. On Windows the gains are a bit less because we are
limited by the poor performance of the DiaSdk for symbol loading.
An upcoming PR will introduce raw_pdb for another massive improvement.